change datatype fld_c to decimal

+++ MVC_SYSTEM/Models/tbl_ByrCarumanTambahan.cs
@@ -68,7 +68,7 @@
 
         public decimal? fld_Q { get; set; }
 
-        public int? fld_C { get; set; }
+        public decimal? fld_C { get; set; }
 
         public decimal? fld_LP { get; set; }
